# Goals

Speed up LS operation on sharded directories [#4908](https://github.com/ipfs/go-ipfs/issues/4908)
Test the use of [EnumerateChildrenAsync](https://github.com/ipfs/go-merkledag/blob/master/merkledag.go#L351) as a fast way of fetching an arbitrary MerkleDAG w/o concern for order (https://github.com/ipfs/go-ipfs/issues/5487)

# Implementation

In the `EnumLinks` method for HAMT, instead of calling ForEachLink (synchronous), use `EnumerateChildrenAsync` 

`EnumerateChildrenAsync` requires a root CID, a `GetLinks` function which returns a list of child IPLD Links to traverse for a given CID, and a `visit` function which returns a boolean of whether to traverse a given CID. In our case, `visit` is a function that just returns true (always visit), and filtering nodes to visit is instead handled by filtering the child links returned by our `getLinks` function. The `getLinks` function is generated by `makeAsyncTrieGetLinks` and does the following:

- Given a CID for a directory shard, fetch the full node for that shard.
- Enumerate the nodes child links and determine which are shards vs shardValues (i.e. actual files) using the name of the link (above a certain length = shard value)
- For each shardValue, call the callback passed to `makeAsyncTrieGetLinks` so that `EnumLinks` can add the value to the links it will return
- Return from the function all of the links that are actual shards so `EnumerateChildrenAsync` will visit them
- `EnumLinks` uses a mutex when appending to the list of links so that appends are thread safe

# For Discussion

In tandem with increasing the [FetchConcurrency](https://github.com/ipfs/go-merkledag/blob/master/merkledag.go#L345) for `EnumerateChildrenAsync` to 128 and adding sessions to the LS operation, this produces the following results for fetching the Wikipedia directory (running `ipfs repo gc && time ipfs ls --resolve-type=false /ipfs/QmT5NvUtoM5nWFfrQdVrFtvGfKFmG7AHE8P34isapyhCxX/wiki/`):

From the internet: `ipfs ls --resolve-type=false   6.08s user 0.86s system 0% cpu 38:30.81 total`
From a machine on local network: `ipfs ls --resolve-type=false   6.68s user 1.13s system 5% cpu 2:17.42 total`

The improvements are simply not as dramatic with the current `FetchConcurrency` of 8. I am also going to experiment with adding `GetMany` calls inside make trie as a way of prefetching children with greater overall concurrency without changing `FetchConcurrency`.

I also wonder if it is neccesary to actually call NewHamtFromDag and whether there are potentially improvements in performance/memory/GC usage by skipping this extra processing step.

I am going to submit a seperate PR for adding sessions to LS, which I think can be merged independently.

Also still need to add tests -- this is not a final implementation