@Stebalien 2018-09-21
For comparison, much time does it take *without* this patch?