@Stebalien 2018-09-21
Np. I'll be traveling 3pm onward anyways.