@hannahhoward 2018-09-24
Follow up on timing:

For internet fetch, I ran `ls` against that sharded directory on Friday and turned it off after it had been running through today without finishing. So major win there.

For local machine, I get the following less than impressive results:

Old Code:
```
ipfs ls --resolve-type=false   4.41s user 0.64s system 1% cpu 4:22.82 total 
```
This PR w/ existing Fetch Concurrency:
```
ipfs ls --resolve-type=false   6.40s user 0.92s system 4% cpu 3:01.85 total 8
```
FetchConcurrency at 256:
```
ipfs ls --resolve-type=false   6.39s user 1.01s system 5% cpu 2:23.57 total 256
```

My intuition here is the speed limit is either the size of the WantList -- since there's only a single peer (I ran this disconnected from the internet) -- or just the CPU (which gets pegged hard in all these cases)