@schomatis 2018-09-29
> @schomatis what was your "over the internet" perf win?

I was replicating @hannahhoward's test from https://github.com/ipfs/go-unixfs/pull/19#issuecomment-423872605

> For internet fetch, I ran `ls` against that sharded directory on Friday and turned it off after it had been running through today without finishing. So major win there.

and on my local machine (with a below average connection) the `ipfs ls` command was downloading the remote content 5-10x faster with this PR (leaving the `FetchGraphConcurrency` at 8).