@hannahhoward 2018-10-02
@schomatis make some readability changes -- see https://github.com/ipfs/go-unixfs/pull/19/commits/c5f90672afa179343289b18535da1ad9bd39d6f3 per your request. Let me know what you think.