@kevina 2018-10-03
I don't want to block this, but I would like to get some compassion to how well this change compares to my more comprehensive solution at https://github.com/ipfs/go-ipfs/pull/4979.  Even if we don't go with that solution it will give us an idea of what we should aim to achieve.