@eingenito 2018-10-03
@kevina I know a little bit about the code so I'll take a shot. I think that after some discussion we relaxed the requirement to return the listing in a stable order. Our reasoning was that readdir(), as an example, doesn't guarantee an repeatable order, and that the retrieval order of entries in the sharded structure is effectively random anyway (although it could be stable). If a sorted output is needed then the results can be sorted after fetching. 

Once that restriction was lifted it meant that we could reuse existing code to do the async walk of the DAG, and that code had some advantages like; it already existed and was more general purpose (dealing with Nodes and []Link), and it didn't bother to wait for all children within a node before recursing which we think (probably) gives us a higher possibility of parallelism in some tree shapes and (quite probably) gives us a quicker time to begin returning _any_ entries. Which is good, but will require additional changes to the API to stream results. 

I think I have that right. People can chime in if I've misunderstood something.