@schomatis 2018-10-03
> @schomatis make some readability changes -- see [45f703f](https://github.com/ipfs/go-unixfs/commit/45f703ffd82511920ec198aa26a7f87e3dfe0366) per your request. Let me know what you think.

Great! The `linkType` and `childLinkType` make a *big* difference in understanding and reasoning about the code.